[#10] Move Dice Library under `Libraries/`
