+++ Source/OpenPF2Core/Private/Abilities/Weapons/PF2WeaponBase.cpp
@@ -12,10 +12,9 @@
 
 #include "Abilities/Weapons/PF2WeaponBase.h"
 
-#include "PF2DiceLibrary.h"
 #include "PF2TemlCalculation.h"
-
 #include "Abilities/PF2CharacterAbilitySystemComponentInterface.h"
+#include "Libraries/PF2DiceLibrary.h"
 
 void UPF2WeaponBase::ActivateAbility(
 	const FGameplayAbilitySpecHandle     Handle,
