+++ Source/OpenPF2Core/Private/Libraries/PF2DiceLibrary.cpp
@@ -3,7 +3,7 @@
 // This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
 // distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-#include "PF2DiceLibrary.h"
+#include "Libraries/PF2DiceLibrary.h"
 
 #include "PF2ArrayUtilities.h"
