+++ Source/OpenPF2Core/Public/Libraries/PF2DiceLibrary.h
